Add lazy install/load support for YouCompleteMe.
The YouCompleteMe plugin is...complicated. And massive. And not always
required. Therefore, this code will not register and install the plugin
unless explicitly asked for by a user. Additionally, the plugin will not
be loaded unless explicitly asked for by a user once installed. In this
way, systems that cannot run this plugin will not clone the repository
(and massive submodule footprint), and will not attempt to run the
post-installation which may require additional software to be installed.

Installation
------------

To install the plugin, a command called `:YcmInstall` is registered
which will register the plugin, perform the installation (i.e. the Git
clone and submodule clones), and execute the `./install.py`
post-installation program with flags computed by checking for language
support locally installed.

Loading
-------

Once installed, the plugin is left in an un-loaded state which means
that the load time of Vim is not impacted, extra forked process won't
consume CPU and RAM, etc. To load the plugin, a command called
`:YcmLoad` is registered which will load the plugin into Vim and leave
your session ready to go.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>