Finally, properly support base16 color scheme.
This took quite a while to *finally* figure out. There were extra colors
to map in `Xresources` for urxvt and Termite worked with a 256-color
aware base16 theme as well.

This final bit was to update the `$TERM` detection to clear the
background color in order to support terminal emulator transparency.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>