Add a spelling strategy.
This was a long time coming. This change enables spell checking by
default (which works by default in concert with syntax highlighting mode
to only work in comments when viewing code) for `en_us` and `en_ca`. A
keyboard shortcut of `<leader> s` toggles spell checking on and off if
it becomes too distracting. Finally, the highlighting of words has been
updated to use font styling and underlining rather than
foreground/background coloring.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>