Auto-enable ALE's fix-on-save behavior.
Note that by default the list of fixers currently only includes using
`shfmt` on shell-bassed files if the program exists. The `<leader>tf`
mapping will toggle the fix-on-save behavior on and off and the
`<leader>sf` mapping can display the status of the toggle. In this way,
a user can disable code formatting on save for codebases that don't yet
use auto formatting.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>