Add some custom Javascript behavior, similar to Rust.
This behavior gets triggerd on the first load of a javascript file to
ensure that ALE is properly set up and that any supporting programs
(currently only `prettier`) is present.

An `:InstallPrettier` command is present if Prettier cannot be found on
the current `$PATH` and this command will update the state of ALE's
fixers upon successful installation.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>