Move `vint` installation detection to `autoload/`
This change moves both the installation and detection of `vim-vint`
under `~/.vim/autload/vint` into `install.vim` and `detect.vim`
respectively. In this way, the detection logic should only be loaded if
required (i.e. lazy) and once at most which is an improvement over
defining the code in `~/.vim/ftplugin` which would load for each new
buffer.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>