Remove YouCompleteMe install/manage logic
This change removes YouCompleteMe from the Vim setup as it has not been
highly used in the practical sense here. Most languages are migrating to
a language server model which tend to support the functionality that
YouCompleteMe was offering, but without the compilation overhead.

If it's ever needed again, well `git revert` is great for that.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>