Assume that ft=sh are coloring POSIX shell code
Surprisingly, Vim takes a very conservative view of the `/bin/sh`
interpreter as being original Bourne shell (i.e. not POSIX-compliant).
Further, in order to determine the parsing mode it will try to resolve
an symlinks to the interpreter after parsing the shebang line, meaning
that Vim's answer could depend on the system which is currently editing
the file.

Being that it's not 42 years ago (really:
https://en.wikipedia.org/wiki/Bourne_shell), let's assume POSIX by
default here. In fact, this solves such an epic mis-highlighting of the
`$(...)` syntax (it is read as an error and colored red) that I thought
that's how it was supposed to color command substitution!

And you know what? I bet I'm going to surprise some other Vim users with
this nugget, you think?

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>