For now, keep `set nocompatible` in `.vimrc`
This ensures that if this Vim setup is sourced in via `vim -u`, it will
still run in nocompatible mode. In other words, no matter how this is
used, the mode will be the same even if 99% of the time nocompatible is
properly detected and set by Vim/GVim itself on startup.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>