Add `vim-markdown-composer` plugin
This is a pretty damn cool plugin which will open the rendered Markdown
in a web browser and keep it in sync to allow for realtime previewing.
The main engine is written in Rust and this commit sets it up lazily to
delay full installation if Rust isn't detected on the system.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>