[rust] Update strategy for the default active target
Due to some updates in the Installation Profiles PR in Rustup
(rust-lang/rustup.rs#1673), as of Rustup 1.20.0, there is no longer a
` (default)` ending indicator on the active target when calling `rustup
target list`. Looking at the update, it's not totally clear that this
was ever completely accurate so another strategy was in order.

This change, uses the knowledge of the current active toolchain (via
`rustup show active-toolchain`) to list all installed components for
that toolchain. The `rustc` component should be installed in an active
toolchain (although this is honestly a stated assumption which could be
invalidated), so if we strip off `rustc-` from the start of the
component name, we should be left with the full target triple of the
active toolchain, hence the "default active target".

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>