Migrate from ALE to coc.nvim
This was a major migration of behavior and setup, but should be a
massive payoff. The motivating factor was to try out `rust-analyzer`
which is already a treat to use. The bonus was all the other linting and
formatting that coc.nvim supports out of the box via its extensions
which are its plugin system.

Chances are that not 100% of behavior is preserved across all filetypes
and languages, but these updates should yield a workable Vim setup to
use as a daily driver.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>