Improve `update.sh` implementation for already installed case
Note that there is still an issue when running the script for the first
time on an installation Vim which has not yet been set up. When coc
starts for the first time, it will start installing extensions but since
that appears to happen asynchronously, there is as of yet no way to tell
that it's finished or to wait until uninstalled extensions are
installed. What a hard problem to solve :/

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>