Show 8-col tabstops in `sh` filetypes for full tab chars
This helps to better display heredocs using the `<<-EOF` style which
supports stripping tab characters from the start each line. In this case
it would be ideal to display the full 8-character width of the tab stop
so as not to start authoring lines that end up being very long once
displayed in simple pagers such as `more`/`less` (even `cat`) and online
source viewers such as GitHub, BitBucket, etc.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>