Set custom `viminfofile` location
This is intended to help in environments where a user may not have full
read/write control over their own home directory but has some control
over the setup. In this case `~/.viminfo` is the default location and a
temporary file `~/.viminfo.tmp` will often get created when trying to
atomically update the main file.

This change updates the location to `~/.vim/cache/viminfo` which will be
a dedicated directory which could be mounted as read/write for this
purpose.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>